1. 添加清除缓存文件接口，以后要添加权限验证； 2. 添加IP定位GPS的示例； 3. 将核心入口文件拆分；

+++ core/common/function.func.php
@@ -175,15 +175,18 @@ function removeFile($path){
 }
 //删除文件夹
 function removeDir($path) {
+    if(!is_dir($path)){
+        return true;
+    }
     //先删除目录下的文件：
     $dh=opendir($path);
     while ($file=readdir($dh)) {
         if($file!="." && $file!="..") {
-            $fullpath=$path."/".$file;
+            $fullpath=$path."/".$file;echo $fullpath;
             if(!is_dir($fullpath)) {
                 unlink($fullpath);
             } else {
-                deldir($fullpath);
+                removeDir($fullpath);
             }
         }
     }
@@ -201,4 +204,34 @@ function getFileContent($url){
     $context = stream_context_create($opts);
     $data = file_get_contents($url, false, $context);
     return  $data;
+}
+
+/**
+ * 获取客户端IP地址
+ * @param integer $type 返回类型 0 返回IP地址 1 返回IPV4地址数字
+ * @param boolean $adv 是否进行高级模式获取（有可能被伪装）
+ * @return mixed
+ */
+function get_client_ip($type = 0,$adv=false) {
+    $type       =  $type ? 1 : 0;
+    static $ip  =   NULL;
+    if ($ip !== NULL) return $ip[$type];
+    if($adv){
+        if (isset($_SERVER['HTTP_X_FORWARDED_FOR'])) {
+            $arr    =   explode(',', $_SERVER['HTTP_X_FORWARDED_FOR']);
+            $pos    =   array_search('unknown',$arr);
+            if(false !== $pos) unset($arr[$pos]);
+            $ip     =   trim($arr[0]);
+        }elseif (isset($_SERVER['HTTP_CLIENT_IP'])) {
+            $ip     =   $_SERVER['HTTP_CLIENT_IP'];
+        }elseif (isset($_SERVER['REMOTE_ADDR'])) {
+            $ip     =   $_SERVER['REMOTE_ADDR'];
+        }
+    }elseif (isset($_SERVER['REMOTE_ADDR'])) {
+        $ip     =   $_SERVER['REMOTE_ADDR'];
+    }
+    // IP地址合法验证
+    $long = sprintf("%u",ip2long($ip));
+    $ip   = $long ? array($ip, $long) : array('0.0.0.0', 0);
+    return $ip[$type];
 }
